[WIP] dead code removal and other refactoring.
Removing dead components and related code:
LockSwitch (w&a)
AllExpectations (scribe)
Divider (w&a)
DisabledAlert (scribe)
ScribeOptions (scribe)
