--- frontend/src/app/components/Divider/Divider.scss
@@ -1,5 +1,0 @@
-div.divider {
-  cursor: col-resize;
-  text-align: center;
-  font-size: 0.5em;
-}
