--- frontend/src/app/components/LockSwitch/LockSwitch.scss
@@ -1,5 +1,0 @@
-.lock-button {
-  i {
-    width: 1em;
-  }
-}
