+++ frontend/src/app/components/Review/Expectations.module.scss
@@ -11,14 +11,3 @@
   border: 0;
   border-radius: 0;
 }
-
-h6 {
-  font-weight: bolder !important;
-  margin-bottom: 0 !important;
-}
-
-nav {
-  .nav-item {
-    font-weight: 600;
-  }
-}
