+++ frontend/vite.config.ts
@@ -17,7 +17,6 @@ export default defineConfig(({ mode }) => ({
         main: resolve(__dirname, 'index.html'),
         deeplink: resolve(__dirname, 'deeplink.html'),
         review: resolve(__dirname, 'review.html'),
-        expectations: resolve(__dirname, 'expectations.html'),
         genlink: resolve(__dirname, 'genlink.html'),
       },
       output: {
